+++ CHANGELOG.md
@@ -25,6 +25,7 @@
     - Getting tree object from projects.
     - Updating a file in a project.
     - Getting merge requests associated with a commit.
+    - Getting commits of a merge request.
   * New parameters for creating projects:
     - `container_registry_access_level`
     - `merge_pipelines_enabled`